+++ src/xyz/oihana/schema/constants/Oihana.php
@@ -3,6 +3,7 @@
 namespace xyz\oihana\schema\constants;
 
 use org\schema\constants\traits\Properties;
+use xyz\oihana\schema\constants\traits\AppointmentsTrait;
 use xyz\oihana\schema\constants\traits\AuditTrait;
 use xyz\oihana\schema\constants\traits\AuthTrait;
 use xyz\oihana\schema\constants\traits\BusinessTrait;
@@ -38,6 +39,7 @@ class Oihana
     public const string SCHEMA = 'https://schema.oihana.xyz';
 
     use Properties        , // All schema.org properties
+        AppointmentsTrait ,
         AuditTrait        ,
         AuthTrait         ,
         BusinessTrait     ,