+++ src/xyz/oihana/schema/constants/traits/appointments/AppointmentTrait.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace xyz\oihana\schema\constants\traits\appointments;
+
+/**
+ * The property name constants of the {@see \xyz\oihana\schema\appointments\Appointment} class.
+ *
+ * @package xyz\oihana\schema\constants\traits\appointments
+ * @author  Marc Alcaraz (eKameleon)
+ * @since   1.5.0
+ */
+trait AppointmentTrait
+{
+    const string APPOINTMENT_STATUS = 'appointmentStatus' ;
+    const string APPOINTMENT_TYPE   = 'appointmentType'   ;
+    const string REPORT             = 'report'            ;
+    const string TAGS               = 'tags'              ;
+}