+++ src/xyz/oihana/schema/constants/traits/appointments/VisitReportTrait.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace xyz\oihana\schema\constants\traits\appointments;
+
+/**
+ * The property name constants of the {@see \xyz\oihana\schema\appointments\VisitReport} class.
+ *
+ * @package xyz\oihana\schema\constants\traits\appointments
+ * @author  Marc Alcaraz (eKameleon)
+ * @since   1.5.0
+ */
+trait VisitReportTrait
+{
+    const string MOOD    = 'mood'    ;
+    const string OUTCOME = 'outcome' ;
+    const string TOPICS  = 'topics'  ;
+}